COMMON: Make inflateZlibInstallShield use gzioreadstream
